requirements.txt

ray[data]>=2.8.0
boto3>=1.26.0
s3fs>=2023.1.0
pyarrow>=10.0.0
pyarrow>=12.0.0
boto3>=1.26.0
pandas>=1.5.0

# SALLMA/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-cohere
langgraph
sqlalchemy
sqlalchemy-cockroachdb
python-dotenv
pandas
psycopg2-binary
